# Get rid of boost::optional leftovers