--- a/include/engine/engine_config.hpp
+++ b/include/engine/engine_config.hpp
@@ -79,7 +79,7 @@ struct EngineConfig final
     int max_locations_map_matching = -1;
     double max_radius_map_matching = -1.0;
     int max_results_nearest = -1;
-    boost::optional<double> default_radius = -1.0;
+    double default_radius = -1.0;
     int max_alternatives = 3; // set an arbitrary upper bound; can be adjusted by user
     bool use_shared_memory = true;
     std::filesystem::path memory_file;