:) 



<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1151.01<br/>plain u32: 1155.63<br/>aliased double: 1199.74<br/>plain double: 1200.37 | aliased u32: 1096.92<br/>plain u32: 1066<br/>aliased double: 951.559<br/>plain double: 951.917 |
| e2e_match_ch | Ops: 45.73 ± 0.06 ops/s. Best: 45.79 ops/s<br/>Total: 2864.80ms ± 3.60ms. Best: 2860.88ms<br/>Min time: 2.19ms ± 0.02ms<br/>Mean time: 21.87ms ± 0.03ms<br/>Median time: 15.68ms ± 0.03ms<br/>95th percentile: 72.01ms ± 0.31ms<br/>99th percentile: 89.15ms ± 1.00ms<br/>Max time: 100.23ms ± 0.33ms | Ops: 45.51 ± 0.17 ops/s. Best: 45.72 ops/s<br/>Total: 2878.22ms ± 10.89ms. Best: 2865.50ms<br/>Min time: 2.14ms ± 0.05ms<br/>Mean time: 21.98ms ± 0.08ms<br/>Median time: 15.86ms ± 0.08ms<br/>95th percentile: 71.90ms ± 0.34ms<br/>99th percentile: 88.86ms ± 0.35ms<br/>Max time: 100.84ms ± 0.29ms |
| e2e_match_mld | Ops: 65.27 ± 0.19 ops/s. Best: 65.48 ops/s<br/>Total: 2006.97ms ± 6.07ms. Best: 2000.48ms<br/>Min time: 1.78ms ± 0.02ms<br/>Mean time: 15.32ms ± 0.05ms<br/>Median time: 8.21ms ± 0.07ms<br/>95th percentile: 50.84ms ± 0.10ms<br/>99th percentile: 58.69ms ± 0.14ms<br/>Max time: 68.31ms ± 0.20ms | Ops: 64.59 ± 0.25 ops/s. Best: 64.87 ops/s<br/>Total: 2028.30ms ± 8.38ms. Best: 2019.49ms<br/>Min time: 1.79ms ± 0.04ms<br/>Mean time: 15.48ms ± 0.06ms<br/>Median time: 8.35ms ± 0.11ms<br/>95th percentile: 51.23ms ± 0.44ms<br/>99th percentile: 59.10ms ± 0.40ms<br/>Max time: 68.44ms ± 0.20ms |
| e2e_nearest_ch | Ops: 804.00 ± 4.82 ops/s. Best: 808.80 ops/s<br/>Total: 1243.76ms ± 7.53ms. Best: 1236.40ms<br/>Min time: 1.06ms ± 0.00ms<br/>Mean time: 1.24ms ± 0.01ms<br/>Median time: 1.16ms ± 0.01ms<br/>95th percentile: 1.64ms ± 0.01ms<br/>99th percentile: 1.69ms ± 0.01ms<br/>Max time: 4.21ms ± 2.29ms | Ops: 795.36 ± 4.69 ops/s. Best: 800.14 ops/s<br/>Total: 1257.38ms ± 8.00ms. Best: 1249.78ms<br/>Min time: 1.07ms ± 0.01ms<br/>Mean time: 1.26ms ± 0.01ms<br/>Median time: 1.17ms ± 0.01ms<br/>95th percentile: 1.66ms ± 0.01ms<br/>99th percentile: 1.71ms ± 0.01ms<br/>Max time: 4.25ms ± 2.34ms |
| e2e_nearest_mld | Ops: 808.23 ± 6.83 ops/s. Best: 816.55 ops/s<br/>Total: 1237.27ms ± 11.10ms. Best: 1224.66ms<br/>Min time: 1.06ms ± 0.01ms<br/>Mean time: 1.24ms ± 0.01ms<br/>Median time: 1.15ms ± 0.01ms<br/>95th percentile: 1.63ms ± 0.01ms<br/>99th percentile: 1.69ms ± 0.02ms<br/>Max time: 4.10ms ± 2.35ms | Ops: 806.81 ± 5.61 ops/s. Best: 814.93 ops/s<br/>Total: 1239.38ms ± 9.12ms. Best: 1227.10ms<br/>Min time: 1.05ms ± 0.01ms<br/>Mean time: 1.24ms ± 0.01ms<br/>Median time: 1.15ms ± 0.01ms<br/>95th percentile: 1.64ms ± 0.01ms<br/>99th percentile: 1.70ms ± 0.01ms<br/>Max time: 4.17ms ± 2.37ms |
| e2e_route_ch | Ops: 353.20 ± 2.18 ops/s. Best: 355.56 ops/s<br/>Total: 2831.91ms ± 19.94ms. Best: 2812.50ms<br/>Min time: 1.27ms ± 0.02ms<br/>Mean time: 2.83ms ± 0.02ms<br/>Median time: 2.84ms ± 0.02ms<br/>95th percentile: 3.76ms ± 0.06ms<br/>99th percentile: 4.22ms ± 0.21ms<br/>Max time: 6.94ms ± 1.71ms | Ops: 344.07 ± 3.87 ops/s. Best: 348.49 ops/s<br/>Total: 2907.76ms ± 33.94ms. Best: 2869.48ms<br/>Min time: 1.29ms ± 0.01ms<br/>Mean time: 2.91ms ± 0.03ms<br/>Median time: 2.92ms ± 0.04ms<br/>95th percentile: 3.84ms ± 0.07ms<br/>99th percentile: 4.25ms ± 0.12ms<br/>Max time: 6.52ms ± 1.80ms |
| e2e_route_mld | Ops: 297.02 ± 0.60 ops/s. Best: 298.08 ops/s<br/>Total: 3366.76ms ± 7.14ms. Best: 3354.83ms<br/>Min time: 1.27ms ± 0.02ms<br/>Mean time: 3.37ms ± 0.01ms<br/>Median time: 3.40ms ± 0.01ms<br/>95th percentile: 4.56ms ± 0.02ms<br/>99th percentile: 5.09ms ± 0.03ms<br/>Max time: 7.19ms ± 1.60ms | Ops: 289.19 ± 1.04 ops/s. Best: 290.80 ops/s<br/>Total: 3458.33ms ± 12.61ms. Best: 3438.80ms<br/>Min time: 1.30ms ± 0.03ms<br/>Mean time: 3.46ms ± 0.01ms<br/>Median time: 3.48ms ± 0.02ms<br/>95th percentile: 4.72ms ± 0.03ms<br/>99th percentile: 5.22ms ± 0.09ms<br/>Max time: 7.46ms ± 1.53ms |
| e2e_table_ch | Ops: 302.00 ± 0.82 ops/s. Best: 302.95 ops/s<br/>Total: 3311.14ms ± 9.05ms. Best: 3300.85ms<br/>Min time: 1.73ms ± 0.00ms<br/>Mean time: 3.31ms ± 0.01ms<br/>Median time: 3.31ms ± 0.01ms<br/>95th percentile: 4.58ms ± 0.02ms<br/>99th percentile: 4.91ms ± 0.04ms<br/>Max time: 7.41ms ± 2.25ms | Ops: 299.02 ± 1.51 ops/s. Best: 301.01 ops/s<br/>Total: 3344.51ms ± 17.40ms. Best: 3322.13ms<br/>Min time: 1.76ms ± 0.02ms<br/>Mean time: 3.34ms ± 0.02ms<br/>Median time: 3.36ms ± 0.02ms<br/>95th percentile: 4.63ms ± 0.03ms<br/>99th percentile: 4.99ms ± 0.04ms<br/>Max time: 7.59ms ± 2.45ms |
| e2e_table_mld | Ops: 112.90 ± 0.18 ops/s. Best: 113.19 ops/s<br/>Total: 8857.55ms ± 14.36ms. Best: 8834.47ms<br/>Min time: 3.59ms ± 0.04ms<br/>Mean time: 8.86ms ± 0.01ms<br/>Median time: 8.81ms ± 0.02ms<br/>95th percentile: 13.51ms ± 0.03ms<br/>99th percentile: 14.40ms ± 0.06ms<br/>Max time: 16.45ms ± 1.44ms | Ops: 111.56 ± 0.26 ops/s. Best: 111.86 ops/s<br/>Total: 8963.85ms ± 21.55ms. Best: 8939.49ms<br/>Min time: 3.64ms ± 0.05ms<br/>Mean time: 8.96ms ± 0.02ms<br/>Median time: 8.94ms ± 0.02ms<br/>95th percentile: 13.71ms ± 0.06ms<br/>99th percentile: 14.58ms ± 0.10ms<br/>Max time: 17.38ms ± 1.42ms |
| e2e_trip_ch | Ops: 98.90 ± 0.88 ops/s. Best: 99.72 ops/s<br/>Total: 10107.73ms ± 89.23ms. Best: 10028.44ms<br/>Min time: 1.51ms ± 0.15ms<br/>Mean time: 10.11ms ± 0.09ms<br/>Median time: 9.59ms ± 0.08ms<br/>95th percentile: 17.92ms ± 0.17ms<br/>99th percentile: 19.33ms ± 0.12ms<br/>Max time: 23.40ms ± 1.94ms | Ops: 96.60 ± 0.90 ops/s. Best: 97.49 ops/s<br/>Total: 10352.95ms ± 97.60ms. Best: 10257.09ms<br/>Min time: 1.57ms ± 0.17ms<br/>Mean time: 10.35ms ± 0.10ms<br/>Median time: 9.89ms ± 0.09ms<br/>95th percentile: 18.40ms ± 0.14ms<br/>99th percentile: 19.90ms ± 0.16ms<br/>Max time: 22.26ms ± 1.70ms |
| e2e_trip_mld | Ops: 59.97 ± 0.20 ops/s. Best: 60.17 ops/s<br/>Total: 16673.58ms ± 55.75ms. Best: 16618.72ms<br/>Min time: 1.61ms ± 0.18ms<br/>Mean time: 16.68ms ± 0.06ms<br/>Median time: 16.25ms ± 0.05ms<br/>95th percentile: 27.42ms ± 0.05ms<br/>99th percentile: 29.11ms ± 0.14ms<br/>Max time: 31.00ms ± 0.48ms | Ops: 58.62 ± 0.15 ops/s. Best: 58.89 ops/s<br/>Total: 17056.58ms ± 46.21ms. Best: 16981.96ms<br/>Min time: 1.67ms ± 0.26ms<br/>Mean time: 17.06ms ± 0.05ms<br/>Median time: 16.62ms ± 0.04ms<br/>95th percentile: 28.01ms ± 0.09ms<br/>99th percentile: 29.69ms ± 0.10ms<br/>Max time: 36.23ms ± 5.17ms |
| json-render | String: 6.71826ms<br/>Stringstream: 10.5533ms<br/>Vector: 6.7972ms | String: 6.84561ms<br/>Stringstream: 10.748ms<br/>Vector: 6.86847ms |
| match_ch | Default radius: <br/>4.48617ms/req at 82 coordinate<br/>0.0547094ms/coordinate<br/>Radius 10m: <br/>15.628ms/req at 82 coordinate<br/>0.190586ms/coordinate | Default radius: <br/>4.47879ms/req at 82 coordinate<br/>0.0546194ms/coordinate<br/>Radius 10m: <br/>15.5643ms/req at 82 coordinate<br/>0.189809ms/coordinate |
| match_mld | Default radius: <br/>3.39581ms/req at 82 coordinate<br/>0.0414123ms/coordinate<br/>Radius 10m: <br/>10.9834ms/req at 82 coordinate<br/>0.133944ms/coordinate | Default radius: <br/>2.97613ms/req at 82 coordinate<br/>0.0362943ms/coordinate<br/>Radius 10m: <br/>10.6916ms/req at 82 coordinate<br/>0.130386ms/coordinate |
| osrm_contract | Time: 96.47s Peak RAM: 201.25MB | Time: 96.37s Peak RAM: 201.75MB |
| osrm_customize | Time: 1.31s Peak RAM: 116.75MB | Time: 1.31s Peak RAM: 116.75MB |
| osrm_extract | Time: 11.88s Peak RAM: 404.60MB | Time: 11.89s Peak RAM: 407.69MB |
| osrm_partition | Time: 1.98s Peak RAM: 133.39MB | Time: 1.98s Peak RAM: 135.14MB |
| packedvector | random write:<br/>std::vector 9916.18 ms<br/>util::packed_vector 74017.2 ms<br/>slowdown: 7.46428<br/>random read:<br/>std::vector 8533.9 ms<br/>util::packed_vector 30825.4 ms<br/>slowdown: 3.61212 | random write:<br/>std::vector 9798.61 ms<br/>util::packed_vector 77737.5 ms<br/>slowdown: 7.93353<br/>random read:<br/>std::vector 8448.45 ms<br/>util::packed_vector 30881.8 ms<br/>slowdown: 3.65533 |
| random_match_ch | 500 matches, default radius<br/>ops: 219.43 ± 1.00 ops/s. best: 220.87ops/s.<br/>total: 259.77 ± 1.20ms. best: 258.07ms.<br/>avg: 4.56 ± 0.02ms<br/>min: 0.16 ± 0.01ms<br/>max: 25.01 ± 0.13ms<br/>p99: 25.01 ± 0.13ms<br/><br/>500 matches, radius=10<br/>ops: 65.63 ± 0.06 ops/s. best: 65.72ops/s.<br/>total: 975.20 ± 0.93ms. best: 973.77ms.<br/>avg: 15.24 ± 0.01ms<br/>min: 0.16 ± 0.00ms<br/>max: 217.64 ± 0.59ms<br/>p99: 217.64 ± 0.59ms<br/><br/>500 matches, radius=20<br/>ops: 15.56 ± 0.03 ops/s. best: 15.61ops/s.<br/>total: 4177.51 ± 8.84ms. best: 4163.67ms.<br/>avg: 64.27 ± 0.14ms<br/>min: 0.31 ± 0.01ms<br/>max: 1129.05 ± 3.88ms<br/>p99: 1129.05 ± 3.88ms | 500 matches, default radius<br/>ops: 215.71 ± 0.83 ops/s. best: 216.65ops/s.<br/>total: 264.25 ± 1.01ms. best: 263.09ms.<br/>avg: 4.64 ± 0.02ms<br/>min: 0.16 ± 0.01ms<br/>max: 25.32 ± 0.10ms<br/>p99: 25.32 ± 0.10ms<br/><br/>500 matches, radius=10<br/>ops: 64.37 ± 0.29 ops/s. best: 64.68ops/s.<br/>total: 994.23 ± 4.52ms. best: 989.42ms.<br/>avg: 15.53 ± 0.07ms<br/>min: 0.16 ± 0.00ms<br/>max: 221.05 ± 1.26ms<br/>p99: 221.05 ± 1.26ms<br/><br/>500 matches, radius=20<br/>ops: 15.31 ± 0.02 ops/s. best: 15.34ops/s.<br/>total: 4245.50 ± 5.21ms. best: 4236.20ms.<br/>avg: 65.32 ± 0.08ms<br/>min: 0.31 ± 0.00ms<br/>max: 1149.43 ± 2.52ms<br/>p99: 1149.43 ± 2.52ms |
| random_match_mld | 500 matches, default radius<br/>ops: 313.98 ± 1.60 ops/s. best: 315.63ops/s.<br/>total: 181.55 ± 0.93ms. best: 180.59ms.<br/>avg: 3.19 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 18.80 ± 0.04ms<br/>p99: 18.80 ± 0.04ms<br/><br/>500 matches, radius=10<br/>ops: 110.91 ± 0.27 ops/s. best: 111.29ops/s.<br/>total: 577.03 ± 1.40ms. best: 575.06ms.<br/>avg: 9.02 ± 0.02ms<br/>min: 0.15 ± 0.00ms<br/>max: 109.56 ± 0.38ms<br/>p99: 109.56 ± 0.38ms<br/><br/>500 matches, radius=20<br/>ops: 22.22 ± 0.03 ops/s. best: 22.25ops/s.<br/>total: 2925.44 ± 3.36ms. best: 2921.73ms.<br/>avg: 45.01 ± 0.05ms<br/>min: 0.20 ± 0.00ms<br/>max: 575.42 ± 0.75ms<br/>p99: 575.42 ± 0.75ms | 500 matches, default radius<br/>ops: 322.10 ± 1.60 ops/s. best: 323.90ops/s.<br/>total: 176.97 ± 0.89ms. best: 175.98ms.<br/>avg: 3.10 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 18.36 ± 0.03ms<br/>p99: 18.36 ± 0.03ms<br/><br/>500 matches, radius=10<br/>ops: 113.94 ± 0.23 ops/s. best: 114.16ops/s.<br/>total: 561.70 ± 1.13ms. best: 560.60ms.<br/>avg: 8.78 ± 0.02ms<br/>min: 0.15 ± 0.00ms<br/>max: 106.09 ± 0.37ms<br/>p99: 106.09 ± 0.37ms<br/><br/>500 matches, radius=20<br/>ops: 22.87 ± 0.04 ops/s. best: 22.95ops/s.<br/>total: 2842.40 ± 5.30ms. best: 2832.65ms.<br/>avg: 43.73 ± 0.08ms<br/>min: 0.20 ± 0.00ms<br/>max: 556.91 ± 1.33ms<br/>p99: 556.91 ± 1.33ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 22776.97 ± 104.82 ops/s. best: 22879.66ops/s.<br/>total: 439.05 ± 2.03ms. best: 437.07ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17618.63 ± 55.08 ops/s. best: 17697.91ops/s.<br/>total: 567.59 ± 1.86ms. best: 565.04ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.01ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14275.65 ± 8.67 ops/s. best: 14287.06ops/s.<br/>total: 700.49 ± 0.43ms. best: 699.93ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 22847.80 ± 135.62 ops/s. best: 22987.06ops/s.<br/>total: 437.70 ± 2.62ms. best: 435.03ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.18 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17836.78 ± 32.26 ops/s. best: 17888.58ops/s.<br/>total: 560.64 ± 1.03ms. best: 559.02ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14222.64 ± 258.32 ops/s. best: 14413.89ops/s.<br/>total: 703.48 ± 13.18ms. best: 693.78ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.21 ± 0.06ms<br/>p99: 0.14 ± 0.01ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 22677.22 ± 51.04 ops/s. best: 22787.30ops/s.<br/>total: 440.97 ± 1.00ms. best: 438.84ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.19 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17669.17 ± 26.38 ops/s. best: 17695.89ops/s.<br/>total: 565.96 ± 0.85ms. best: 565.10ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14291.41 ± 21.03 ops/s. best: 14317.91ops/s.<br/>total: 699.72 ± 1.07ms. best: 698.43ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 22931.94 ± 44.82 ops/s. best: 22989.87ops/s.<br/>total: 436.08 ± 0.85ms. best: 434.97ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.03ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17790.33 ± 76.78 ops/s. best: 17866.41ops/s.<br/>total: 562.12 ± 2.43ms. best: 559.71ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.16 ± 0.00ms<br/>p99: 0.12 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14302.36 ± 53.11 ops/s. best: 14367.17ops/s.<br/>total: 699.20 ± 2.60ms. best: 696.03ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.00ms<br/>p99: 0.14 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 502.04 ± 1.89 ops/s. best: 505.27ops/s.<br/>total: 1960.03 ± 7.36ms. best: 1947.46ms.<br/>avg: 1.99 ± 0.01ms<br/>min: 0.33 ± 0.01ms<br/>max: 3.41 ± 0.15ms<br/>p99: 2.90 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 590.88 ± 3.57 ops/s. best: 594.26ops/s.<br/>total: 1692.48 ± 10.31ms. best: 1682.76ms.<br/>avg: 1.69 ± 0.01ms<br/>min: 0.05 ± 0.00ms<br/>max: 4.08 ± 0.08ms<br/>p99: 3.67 ± 0.06ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 991.60 ± 10.16 ops/s. best: 1004.61ops/s.<br/>total: 992.47 ± 10.23ms. best: 979.48ms.<br/>avg: 1.01 ± 0.01ms<br/>min: 0.28 ± 0.00ms<br/>max: 1.65 ± 0.03ms<br/>p99: 1.41 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1049.69 ± 21.22 ops/s. best: 1098.93ops/s.<br/>total: 953.18 ± 18.78ms. best: 909.98ms.<br/>avg: 0.95 ± 0.02ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.47 ± 0.06ms<br/>p99: 2.19 ± 0.04ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 506.10 ± 3.27 ops/s. best: 509.98ops/s.<br/>total: 1944.39 ± 12.63ms. best: 1929.49ms.<br/>avg: 1.98 ± 0.01ms<br/>min: 0.33 ± 0.00ms<br/>max: 3.62 ± 0.49ms<br/>p99: 2.91 ± 0.11ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 576.61 ± 6.90 ops/s. best: 591.32ops/s.<br/>total: 1734.58 ± 19.30ms. best: 1691.12ms.<br/>avg: 1.73 ± 0.02ms<br/>min: 0.05 ± 0.00ms<br/>max: 4.28 ± 0.09ms<br/>p99: 3.77 ± 0.06ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 987.61 ± 8.89 ops/s. best: 1004.73ops/s.<br/>total: 996.45 ± 8.92ms. best: 979.37ms.<br/>avg: 1.01 ± 0.01ms<br/>min: 0.28 ± 0.00ms<br/>max: 1.72 ± 0.10ms<br/>p99: 1.44 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1093.69 ± 11.52 ops/s. best: 1107.29ops/s.<br/>total: 914.47 ± 9.73ms. best: 903.10ms.<br/>avg: 0.91 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.43 ± 0.04ms<br/>p99: 2.09 ± 0.03ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 226.68 ± 0.56 ops/s. best: 227.49ops/s.<br/>total: 4340.99 ± 10.59ms. best: 4325.47ms.<br/>avg: 4.41 ± 0.01ms<br/>min: 0.32 ± 0.00ms<br/>max: 10.14 ± 0.52ms<br/>p99: 7.55 ± 0.07ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 225.61 ± 1.56 ops/s. best: 227.65ops/s.<br/>total: 4432.77 ± 30.82ms. best: 4392.62ms.<br/>avg: 4.43 ± 0.03ms<br/>min: 0.05 ± 0.00ms<br/>max: 10.08 ± 0.20ms<br/>p99: 8.92 ± 0.14ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 285.40 ± 0.89 ops/s. best: 286.51ops/s.<br/>total: 3447.86 ± 10.75ms. best: 3434.49ms.<br/>avg: 3.50 ± 0.01ms<br/>min: 0.31 ± 0.01ms<br/>max: 8.77 ± 0.22ms<br/>p99: 6.13 ± 0.11ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 280.93 ± 2.93 ops/s. best: 284.30ops/s.<br/>total: 3560.15 ± 37.53ms. best: 3517.44ms.<br/>avg: 3.56 ± 0.04ms<br/>min: 0.04 ± 0.00ms<br/>max: 8.02 ± 0.41ms<br/>p99: 6.99 ± 0.08ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 242.56 ± 2.43 ops/s. best: 245.19ops/s.<br/>total: 4057.16 ± 40.82ms. best: 4013.22ms.<br/>avg: 4.12 ± 0.04ms<br/>min: 0.33 ± 0.00ms<br/>max: 9.01 ± 0.16ms<br/>p99: 7.04 ± 0.18ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 236.74 ± 2.10 ops/s. best: 240.21ops/s.<br/>total: 4224.44 ± 37.38ms. best: 4163.00ms.<br/>avg: 4.22 ± 0.04ms<br/>min: 0.05 ± 0.00ms<br/>max: 10.19 ± 0.63ms<br/>p99: 8.63 ± 0.10ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 323.05 ± 4.71 ops/s. best: 332.27ops/s.<br/>total: 3046.88 ± 44.37ms. best: 2961.49ms.<br/>avg: 3.10 ± 0.05ms<br/>min: 0.31 ± 0.00ms<br/>max: 7.29 ± 0.16ms<br/>p99: 5.50 ± 0.12ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 297.09 ± 1.88 ops/s. best: 299.39ops/s.<br/>total: 3366.13 ± 21.31ms. best: 3340.18ms.<br/>avg: 3.37 ± 0.02ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.92 ± 0.43ms<br/>p99: 6.65 ± 0.08ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1388.78 ± 10.34 ops/s. best: 1396.98ops/s.<br/>total: 180.03 ± 1.36ms. best: 178.96ms.<br/>avg: 0.72 ± 0.01ms<br/>min: 0.52 ± 0.00ms<br/>max: 1.06 ± 0.29ms<br/>p99: 0.89 ± 0.06ms<br/><br/>250 tables, 25 coordinates<br/>ops: 162.02 ± 0.46 ops/s. best: 162.38ops/s.<br/>total: 1543.06 ± 4.36ms. best: 1539.59ms.<br/>avg: 6.17 ± 0.02ms<br/>min: 5.62 ± 0.04ms<br/>max: 7.03 ± 0.34ms<br/>p99: 6.69 ± 0.03ms<br/><br/>250 tables, 50 coordinates<br/>ops: 79.67 ± 0.18 ops/s. best: 79.86ops/s.<br/>total: 3137.95 ± 7.08ms. best: 3130.52ms.<br/>avg: 12.55 ± 0.03ms<br/>min: 11.83 ± 0.04ms<br/>max: 13.51 ± 0.17ms<br/>p99: 13.36 ± 0.12ms | 250 tables, 3 coordinates<br/>ops: 1405.07 ± 9.56 ops/s. best: 1412.97ops/s.<br/>total: 177.94 ± 1.22ms. best: 176.93ms.<br/>avg: 0.71 ± 0.00ms<br/>min: 0.52 ± 0.00ms<br/>max: 1.04 ± 0.29ms<br/>p99: 0.88 ± 0.05ms<br/><br/>250 tables, 25 coordinates<br/>ops: 163.68 ± 0.28 ops/s. best: 164.08ops/s.<br/>total: 1527.40 ± 2.60ms. best: 1523.65ms.<br/>avg: 6.11 ± 0.01ms<br/>min: 5.54 ± 0.04ms<br/>max: 6.70 ± 0.07ms<br/>p99: 6.60 ± 0.03ms<br/><br/>250 tables, 50 coordinates<br/>ops: 79.92 ± 0.10 ops/s. best: 80.10ops/s.<br/>total: 3127.98 ± 3.79ms. best: 3121.15ms.<br/>avg: 12.51 ± 0.02ms<br/>min: 11.80 ± 0.02ms<br/>max: 14.01 ± 0.92ms<br/>p99: 13.27 ± 0.06ms |
…