@DennisOSRM 2024-07-02
I am traveling this week. The plan is to get to it the next couple of days.